+++ sw/device/tests/sim_dv/alert_handler_lpg_sleep_mode_alerts.c
@@ -321,7 +321,7 @@ static void execute_test_phases(uint8_t test_phase, uint32_t ping_timeout_cyc) {
 
     // Set the AON timer to send a wakeup signal in ~10-50us.
     CHECK_STATUS_OK(aon_timer_testutils_wakeup_config(
-        &aon_timer, rand_testutils_gen32_range(2, 10)));
+        &aon_timer, (uint32_t)rand_testutils_gen32_range(2, 10)));
 
     // Trigger the SV side to inject fault.
     // DO NOT CHANGE THIS: it is used to notify the SV side.
@@ -375,7 +375,7 @@ static void execute_test_phases(uint8_t test_phase, uint32_t ping_timeout_cyc) {
 
     // Set the AON timer to send a wakeup signal in ~10-50us.
     CHECK_STATUS_OK(aon_timer_testutils_wakeup_config(
-        &aon_timer, rand_testutils_gen32_range(2, 10)));
+        &aon_timer, (uint32_t)rand_testutils_gen32_range(2, 10)));
     // Enter the normal sleep mode
     enter_low_power(/*deep_sleep=*/false);
   } else {
